1.9.9 Update AO Depth Map Precision Correction
This update the loss of Precision in the Depth map when AO was turned on. Now there should not be any Precision loss.

Also added more control and Adjustments to 3D AO.

Please note this is a Adjusted AO for the depth map that gives better 3D seperation between objects and it only effects the Depth Buffer.