SuperDepth3D Major Updates
SuperDepth3D
+Convergence Adjustment Equation Added.
+Convergence Code Simplification.
-Balance code removed.
+Depth Boost was moved.
SuperDepth3D_FB
+ Convergence Adjustment Equation Added.
+Convergence Code Simplification.
+Debanding Code Added.
-Dithering Code removed.
-Balance code removed.
+Depth Boost was moved.

Notes: Balance was a stop gap for the convergence code.  Default balance is now set too 0.548125. The reason I picked this number is because this what seemed to be good after testing many diffrent games. It, should be set to 0.5 But, I didn't feel right at that number. The previous default was 0.5625. I will be putting a way to change this in the shader later. So this is kind of a testing stage.