Update SuperDepth3D_2.0.0_Next.fx
Defaults changes.
Divergence Max is now 75 with a default of 37.5.
Bug where edges start to sparkle, had to get rid of a few pixies and there pixie magic.
Code Compression, need to keep things tiddy.
Technique renameing so that it don't enable with SuperDepth3D is in the same folder.