SuperDepth3D 2.0.0 Next
Fixed Loop.

Replaced Flow Controll
break; // We have hit the limit so we dont need to loop anymore
With
continue; // Once we hit the limit skip the rest of the loop and go back to thes start of the loop.

This Should Fix the bug that happens in some game where the loopjust hults.