Update Smart_Sharp.fx
General Motion-Based Detection Added.

This will adjust sharpening strength by the user-selected amount. Based on a multiplier that the user selects as well.

This is only the first part of the motion-based sharpening effects that will be added to smart sharp.

As always all changes will be added to NLM_Sharp :knife:  at a later date.