SuperDepth3D and Overwatch Update
SuperDepth3D Update

Fixed the issue with Cusor Scaling in Line Interlaced mode.

Fixed the issue with not displaying the network play warning.

Fixed the issue naming of Reposition Depth.

Changed Depth Postion sligthly so it's from the -1.0 to 1.0 range and not the -4096 to 4096 range.

Overwatch Game Profiles added
+Remnant: From the Ashes
+Vanquish

Game Help
Added information for Vanquish -DGVoodoo2. Along with website.