Update SuperDepth3D_WoWvx.fx
Changed how the Singnag code is Positioned. Also gave the user a method to position the code upper left and lower left of the screen. In the hopes this helps.

In the next update I wil take in to account VSR. So larger then 1080p resoultions can be used and downscaled.