Dimension_Plus Added
+Dimension_Plus
Lets you Shift postion of the screen to give a better sence of depth. Limitations where set in place so that you don't get motion sick. But, keep in mind this is still a risk.

+TobiiEye_FreePie_Depth3D.py
The scrip for eye tracking was made using the TobiiEye c4.

A other tracking method will be added later for now this is what is provied.

If you made your own script/software please feel free to tell me so I can link it  so that others can enjoy.

More Eye / Head tracking shaders will be coming later this year. As well as updates to older shaders like BloomingHDR so you can look at the sun and be like, "My Eyes!"