MULTI Update
SuperDepth3D 2.5.2

Added Specialized Depth Triggers
Specialized depth triggers will allow for real time depth buffer shifting. This is hardcoded for now since I seen this issue in only one game this helps correct it.

Code Compression and Changes to PreDepth not it's a Array output of 3x3 this one done to add more mod depth outputs that can be used though out the shader.

SuperDepth3D_VR+ v2.4.2

Added Specialized Depth Triggers
Specialized depth triggers will allow for real time depth buffer shifting. This is hardcoded for now since I seen this issue in only one game this helps correct it.

Code Compression and Changes to PreDepth not it's a Array output of 3x3 this one done to add more mod depth outputs that can be used though out the shader.

Overwatch Updated 1.9.3

Added Specialized Depth Trigger
This is so that Overwatch can enable this mode.

Profiles Added

+Star Trek EliteForce II
+Dead or Alive 6
+Dead Rising
+Dead Rising 2
+Dead Rising 2 Off The Record
+Dead Rising 3
+Dead Rising 4
+Iron Harvest
+Senran Kagura Peach Ball

Small note about Dead Rising the cam is all over the place in cutscenes so Eye Strain Warning on them.

Senran Kagura Peach Ball was the reason for the new Trigger since this game shifts the depth buffer when in game many times.. But, now the game has Depth.

Game Help Updated

Added information for Star Trek Elite Force 2.
Added information for Dead or Alive 6
Updated information on Dead Rising 3

I Need to port this to github pages..........

As a Side note Dead Rising 3 and 4 will not work in the newer version of reshade. use 4.4. But, this issue will get fixed. Just not soon.

As Always Please Report Bugs. This is a big change. So there may be bugs..........