SuperDepth3D Update v2.7.6
All View Modes added. Total of 6 with Low Med and High for 5 of them.
Adaptive the only one with no Low Med High setting due to it setting it in real time depending on content in the game. Much faster then before.

Please Note Normal Alpha is a the same one and never changes from release to release and is a default View mode. It Has lax setting so it can be used with not so matching content.

Normal Beta is a WIP View mode that is sensitive and keeps more information than Normal Alpha. This is only useful in game where the depth buffer is a exact match to the depth buffer. like in the game "Remnant: From the Ashes."

View Mode Flashback is a old view mode that was used in older versions of SuperDepth3D. It has a glassy look to it. It used to be called Alpha in v2.5.2

View Mode Reiteration is the same thing as Alpha with multi break points.

View Mode Stamped Is the real name for the effect called alpha that was added around 2.6.9.

Confusing I know. That is why I am not using Alpha any more for the view mode as a name.  I only use it an A B C / Alpha Beta Gamma as a letter system.

So once VM1 gets a name or gets phased out Alpha word will be dropped.

VM0 Normal Alpha
VM1 Normal Beta
VM2 FlashBack
VM3 Reiteration
VM4 Stamped
VM5 Adaptive

No update for The VR shader until this gets sorted.

OverWatch Update 2.1.6

Updated
+Mirror's Edge
+Tekken 7
+DarkSiders III
+Crash Bandicoot N.Saine Trilogy
+Control DX12
+Strange Brigade DX12 & Vulkan
+Sekiro Shadows Die Twice
+A Plague Tale Innocence
+Assassin's Creed Odyssey
+Need For Speed: Heat
+Zombie Army Trilogy

New Profile
+Need For Speed: Payback

Games like FH5 and the new FC where checked.

No Depth in FH5 at least not yet.

As for the new FC please read here.

https://github.com/BlueSkyDefender/Depth3D/issues/52#issuecomment-966782695