Update for Dying Light 2
Game settings and mod information Updated please read the file Read Help to make sure this game works as it should. It also has links and information for the mod. Also information on how to disable the Shader option if you Use the Mod.

One issue that I was not able to fix was the player him self floats of the ground.......
Ignore captain hover boots.