Multi Update
SuperDepth3D and VR+  Shaders

Small Change to Auto ZPD. Now it's slightly more performant.
+0.1ms Not much.

Compatibility Power can now got from -1 - 1 and zero is off. The new feature is that if you go negative it emphasizes the offset on  the other area away from the center. The helps with distortions like in the game Kena Bridge of spirits.

SuperDepth3D
Now Defaults to Performant Mode.

SuperDepth3D_VR+
Now Defaults to Performant + VRS Mode.

The reason for this is because 4k in modern games even on a 2080ti is hard. This should help with users complaining about performance.

Also remember to set reshades own performance mode to on.

Overwatch Profiles.
+DreadOut 2
+Sessions Skate Sim
+Destroy All Humans 2 - Reprobed
+What Remains of Edith Finch

Profiles Update
+SpaceHulk DeathWing Enhanced Edition
+Kena Bridge of Spirits
+Call of the Sea