Depth3D Shaders Bug Fixes and Profiles.
SuperDepth3D and SD3DVR+

Updates and Bug fixes:

1. When weapon hand was out of range it reverted back to 1 in depth. This is no longer the case.

2. Fast Trigger Mode Now activates as it should at the level it is requested. This was broken for Everspace 2 and Trackmania.

3.Auto ZPD now uses less instructions to do the same thing.

4.Fixed bug where Fast Trigger mode was defaulting to on when set to zero.

5.Min Level Trigger was not activating when set to the level it should activate. This was re-written. So it able to activate from ZPD Detection Level 0-4 and or ZPD Detection level > #.

6. De_Artifacting Y In the negative range now Reverse the Depth Masking. So it does something now.

7.Removed unnecessary code that just sat there and did nothing.

8. Added Notes as reminders for future changes.

SuperDepth3D Experimental Update

+HDR Compatible Mode Added

Needs to be user activated. Because right now it seems like some game do not report when they are in HDR mode. Because of that this needed to be set by the user for now.

This a preprocessor at the bottom of the main shader. This was not added to the VR one due to lack of HDR VR screens.

Right now this works by normalizing BT2020_To_BT709 and from when expanding it back again.

Only tested this on 3 Games. Needs more user testing.

Overwatch Profiles

+Gripper
+Dredge
+Crab Champions
+Lost in Random
+Fe
+EverSpace 2
+Alien Cube
+CARNAL
+Bloodhound
+Finding the Soul Orb

Updated Profiles

+System ReShock
+Layers of Fear 2023