SuperDepth3D Update
Depth3D Shaders

Added a way to tell if the profile is loaded along with version of the files in side the Shader under Depth3D Information.

Overwatch

Profiles updated
+Lords of the Fallen 2023
Added more UI Detection for multi Windows at resolutions 1080p / 1440p / 4k and 1620p

Also added more information in the Game Help.txt for the Depth add-on Mod and where to get it.

Also adjusted the Cut scene offset.

Keep in mind you need to make sure other shaders do not mess with the Detectors.