Game Pad Update
Added support for the Xinp GamePad add-on: GitHub – BlueSkyDefender/Xinp

SuperDepth3D now features adjustable 3D depth when zooming in, seamlessly integrating with the shader’s existing mouse zoom functionality.

This enhancement will be applied to all profiles utilizing this feature. Please note that the Xinp add-on must be installed to enable this functionality. A ReShade installer version with the add-on included will be released at a later date.

Additional Alpha UI integration has been implemented for both this feature and Wukong.